Moved Spray files and removed original comments
