+++ serverDemo/src/main/scala/games/demo/server/Boot.scala
@@ -1,6 +1,6 @@
-package com.example
+package games.demo.server
 
-import akka.actor.{ActorSystem, Props}
+import akka.actor.{ ActorSystem, Props }
 import akka.io.IO
 import spray.can.Http
 import akka.pattern.ask
@@ -8,14 +8,10 @@ import akka.util.Timeout
 import scala.concurrent.duration._
 
 object Boot extends App {
-
-  // we need an ActorSystem to host our application in
   implicit val system = ActorSystem("on-spray-can")
 
-  // create and start our service actor
   val service = system.actorOf(Props[MyServiceActor], "demo-service")
 
   implicit val timeout = Timeout(5.seconds)
-  // start a new HTTP server on port 8080 with our service actor as the handler
   IO(Http) ? Http.Bind(service, interface = "::0", port = 8080)
 }
